campus-nest/ML/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
groq==0.32.0
PyMuPDF==1.26.5
pymongo==4.15.3
requests==2.32.5
python-dotenv==1.1.1
gunicorn==23.0.0